SortOrderSettingsを独立させた
