+++ Assets/App/Scripts/Utility/UI/SortOrderSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7951923224c36481b8fdbcd7421599ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
